added new apple
finished everything up to GUI. Added a green apple that drops after 5 red apples. I left the layer as default so that the collision of the tree and green apple would send the apple at an angle sometimes or make it faster. This will make the green one harder to get and worth more points in the future. Future ideas to add are cats and basket balls for more or deducted points