Mouse controls basket
Last step I completed was having the mouse control the basket.

Was all going well until I saw that there is a fourth basket higher up than the three that are supposed to be there. It is also controlled by my mouse and interacts with apples as it should. Looking into fixing that now.